transfered delete option from charcard to characterlist
